alpha 0.0.11
IMPLEMENTATIONS:
- make tooltip a scriptable object, thus being applied in all scenes
- there is a tooltip for the HP and MP bars
- tooltips now change in real time, updating while the player hovers over

BUG FIXES:
- spam attacks while targeting

CHANGES:
- status are now using int instead of float

+++ Assets/Scripts/AudioManager.cs
@@ -51,8 +51,11 @@ public class AudioManager : MonoBehaviour
     private void Update()
     {
 
-        if (player != null){
-            ambienceOutput.pitch = Mathf.Lerp(2.5f, .5f, player.getHP() / player.getMaxHP());
+        // adjjusts the player heartbeat sound based on the player's HP
+        // while player is in combat
+        if (player != null)
+        {
+            ambienceOutput.pitch = Mathf.Lerp(2.5f, .5f, (float)player.getHP() / (float)player.getMaxHP());
         }
     }
